requirements.txt
flask
python-dotenv
python-jose
flask-cors
six
azure-storage-blob
authlib
openai
pydub
ffmpeg
srt
flask-socketio
datetime
firebase-admin
nltk
stripe>=5.0.0
auth0-python
azure-functions
azure-functions-durable
gunicorn
gevent>=1.4
azure-ai-textanalytics==5.2.0
azure-ai-translation-document==1.0.0
